requirements.txt
absl-py==2.0.0
accelerate==0.23.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
av==10.0.0
beautifulsoup4==4.12.2
bitsandbytes
blinker==1.6.3
braceexpand==0.1.7
cachetools==5.3.1
certifi==2023.7.22
cffi
charset-normalizer==3.3.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
coloredlogs==15.0.1
configparser==5.3.0
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.3
decorator==4.4.2
decord==0.6.0
deepspeed==0.11.1
diffusers==0.21.4
diffusions==0.1.13
dm-tree==0.1.8
docker-pycreds==0.4.0
dominate==2.8.0
easydict==1.10
einops==0.4.1
exceptiongroup==1.1.3
fastapi==0.104.0
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.12.4
fire==0.5.0
Flask==3.0.0
flatbuffers==1.12
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
ftfy==6.1.1
func-timeout==4.3.5
future
fvcore==0.1.5.post20221221
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==3.50.0
gradio_client==0.6.1
grpcio==1.54.2
h11==0.14.0
h5py==3.10.0
hjson==3.1.0
HTML4Vision==0.4.3
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.6.0
importlib-resources==6.1.0
invisible-watermark==0.2.0
iopath==0.1.10
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==3.1.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
libclang 
libretranslatepy==2.1.1
lightning-utilities==0.9.0
lit==16.0.6
lmdb==1.3.0
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mmcv==2.0.1
mmcv-full==1.7.1
mmengine==0.7.3
model-index==0.1.11
moviepy==1.0.3
mpi4py
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
natsort==8.4.0
netron==7.3.3
networkx==3.1
ninja==1.11.1.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-dali-cuda110==1.32.0
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
octo-rpc==0.4.7
omegaconf==2.3.0
onnx==1.15.0
onnx-simplifier==0.4.0
onnxconverter-common==1.13.0
onnxmltools==1.11.1
onnxruntime==1.11.1
onnxsim-no-ort==0.4.0
open-clip-torch==2.23.0
openai-clip==1.0.1
opencv-python==4.8.1.78
openmim==0.3.7
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.9.9
packaging==23.2
pandas==2.0.2
parso==0.8.3
pathtools==0.1.2
Pillow==9.5.0
pip-check==2.8.1
ply==3.11
portalocker==2.7.0
proglog==0.1.10
protobuf==3.20.1
psutil==5.9.6
pudb==2023.1
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser
pydantic==1.10.13
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
pyre-extensions==0.0.23
PySocks==1.7.1
python-cat==0.0.11
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.1.0
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.7
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.33.1
setproctitle==1.3.2
six==1.16.0
skl2onnx==1.11.2
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
streamlit==1.28.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboardX==2.6
tensorflow-estimator==2.7.0
tensorflow-gpu 
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminaltables==3.1.10
tf2onnx==1.11.1
threadpoolctl==3.2.0
thrift==0.16.0
thriftpy2==0.4.17
tifffile==2023.9.26
timm==0.9.7
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchdata==0.4.1
torchmetrics==1.2.0
torchvision==0.13.1+cu113
tornado==6.3.3
tqdm==4.66.1
transformers==4.34.0
translate==3.6.1
triton==2.1.0
TritonServerClient==0.0.7
typing-inspect==0.9.0
typing_extensions
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.6
urwid==2.2.3
urwid-readline==0.13
uvicorn==0.23.2
validators==0.22.0
wandb==0.15.4
watchdog==3.0.0
wcwidth==0.2.8
webdataset==0.2.48
websockets==11.0.3
Werkzeug==3.0.0
wrapt==1.15.0
xformers==0.0.16
yacs==0.1.8
yapf==0.33.0
yarl==1.9.2
zipp==3.17.0
